Refactor to use withRequest HOC
